requirements.txt
pandas>=1.5.0
numpy>=1.21.0
torch
matplotlib>=3.5.0
requests>=2.28.0
python-dotenv>=0.19.0
onnxruntime
onnxconverter-common
onnxmltools
